Beginning to code some things! We've got the basic way to fetch out very standardized sprites and compose animations from them using a sprite map file. I don't know anything about transparency so that was not included and should be.
The idea is as follows:
A sprite map in this project should be a series of sprites organized in a very regular manner. Each sprite should be of the same width and height, though this space can include null space (that way the artist isn't restricted on size, but the programmer can easily extract the images). The key assumption that I make in my code, however, is that each row of sprites in the map is one "animation path." An "animation path" is the series of frames that are required to show that a certain action occurred/is occurring. In example, if a person were running to the right, that would require some number of frames (we're using 5 or 6) that would loop. One full iteration of this animation is an "animation path." This way, we can just repeat the same path, and when we want to animate a certain action, we can just call the whole path and start from the front.
The SpriteMap class will contain a BufferedImage object (the sprite map image) and an int explaining the height and width of ONE sprite from the map. The SpriteMap class generates AnimationPath objects.
The AnimationPath class will contain a LinkedList of BufferedImage objects. Each of these objects is a